+++ js/main.js
@@ -139,8 +139,8 @@ import {
 
 (()=>{
 "use strict";
-const GAME_VERSION="1.15.26";
-const GAME_RELEASE_NAME="Festungshof-Skalierung";
+const GAME_VERSION="1.15.27";
+const GAME_RELEASE_NAME="Innerer Mauerring-Ausbau";
 const AUTOSAVE_INTERVAL_MS=60_000;
 const discoveredEnemies=loadDiscoveredEnemies();
 function discoverEnemy(type){